pull from upstream

+++ idea/src/org/jetbrains/kotlin/idea/quickfix/RemovePsiElementSimpleFix.kt
@@ -20,11 +20,10 @@ import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.diagnostics.Diagnostic
+import org.jetbrains.kotlin.idea.intentions.RemoveExplicitTypeIntention
+import org.jetbrains.kotlin.idea.util.CommentSaver
 import org.jetbrains.kotlin.lexer.KtTokens
-import org.jetbrains.kotlin.psi.KtFile
-import org.jetbrains.kotlin.psi.KtImportDirective
-import org.jetbrains.kotlin.psi.KtProperty
-import org.jetbrains.kotlin.psi.KtTypeArgumentList
+import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.psi.psiUtil.getNonStrictParentOfType
 
 open class RemovePsiElementSimpleFix(element: PsiElement, private val text: String) : KotlinQuickFixAction<PsiElement>(element) {
@@ -33,7 +32,7 @@ open class RemovePsiElementSimpleFix(element: PsiElement, private val text: Stri
     override fun getText() = text
 
     public override fun invoke(project: Project, editor: Editor?, file: KtFile) {
-        element.delete()
+        element?.delete()
     }
 
     object RemoveImportFactory : KotlinSingleIntentionActionFactory() {
@@ -62,13 +61,15 @@ open class RemovePsiElementSimpleFix(element: PsiElement, private val text: Stri
     object RemoveVariableFactory : KotlinSingleIntentionActionFactory() {
         public override fun createAction(diagnostic: Diagnostic): KotlinQuickFixAction<PsiElement>? {
             val expression = diagnostic.psiElement.getNonStrictParentOfType<KtProperty>() ?: return null
+            if (!RemoveExplicitTypeIntention.redundantTypeSpecification(expression, expression.initializer)) return null
             return object : RemovePsiElementSimpleFix(expression, "Remove variable '${expression.name}'") {
                 override fun invoke(project: Project, editor: Editor?, file: KtFile) {
                     val initializer = expression.initializer
-                    if (initializer != null) {
-                        expression.replace(initializer)
-                    }
-                    else {
+                    if (initializer != null && initializer !is KtConstantExpression) {
+                        val commentSaver = CommentSaver(expression)
+                        val replaced = expression.replace(initializer)
+                        commentSaver.restore(replaced)
+                    } else {
                         expression.delete()
                     }
                 }
